perf(test): support fractional seconds for collection interval
Change --take-time(-i) flag type from int to float64 to allow
sub-second collection intervals such as 0.03 (30ms) or 0.5 (500ms).

Modified files:
- flag.go: change takeTime variable and FlagArgs.TakeTime field from int to float64
- flag.go: replace flag.IntVar with flag.Float64Var for float flag registration
- main.go: adapt waiting prompt and time.Sleep duration calculation for float type
- count.go: adapt time.NewTicker interval calculation for float type

Now users can specify arbitrary decimal seconds for more flexible
metrics collection in high-frequency performance testing scenarios.

+++ test/count.go
@@ -13,7 +13,7 @@ import (
 // Couter collects memory, GC, CPU and pool metrics at a fixed interval.
 func Couter(args FlagArgs, p *agilepool.Pool, f *os.File, start time.Time) {
 	var memStats runtime.MemStats
-	tick := time.NewTicker(time.Duration(args.TakeTime) * time.Second)
+	tick := time.NewTicker(time.Duration(args.TakeTime * float64(time.Second)))
 	defer tick.Stop()
 
 	headerWritten := false
